Wait for open connection when starting hive metastore and server2

+++ actions/hdp/lifecycle.coffee
@@ -110,44 +110,64 @@ lifecyle = module.exports =
       code_skipped: 1
     , (err, started) ->
       callback err, started
+  hive_metastore_status: (ctx, callback) ->
+    # {hive_metastore_pid} = ctx.config.hdp
+    lifecyle.is_pidfile_running ctx, "/var/run/hive/metastore.pid", (err, running) ->
+      callback err, running
   hive_metastore_start: (ctx, callback) ->
-    {hive_user, hive_log_dir, hive_pid_dir} = ctx.config.hdp
-    ctx.execute
-      # su -l hive -c 'nohup hive --service metastore >/var/log/hive/hive.out 2>/var/log/hive/hive.log & echo $! >/var/run/hive/metastore.pid'
-      # su -l hive -c 'hive --service metastore'
-      cmd: "su -l #{hive_user} -c 'nohup hive --service metastore >#{hive_log_dir}/hive.out 2>#{hive_log_dir}/hive.log & echo $! > /var/run/hive/metastore.pid'"
-      code_skipped: 1
-    , (err, started) ->
-      callback err, started
+    {hive_user, hive_log_dir, hive_pid_dir, hive_metastore_host, hive_metastore_port, hive_metastore_timeout} = ctx.config.hdp
+    lifecyle.hive_metastore_status ctx, (err, running) ->
+      return callback err, false if err or running
+      ctx.execute
+        # su -l hive -c 'nohup hive --service metastore >/var/log/hive/hive.out 2>/var/log/hive/hive.log & echo $! >/var/run/hive/metastore.pid'
+        # su -l hive -c 'hive --service metastore'
+        cmd: "su -l #{hive_user} -c 'nohup hive --service metastore >#{hive_log_dir}/hive.out 2>#{hive_log_dir}/hive.log & echo $! > /var/run/hive/metastore.pid'"
+        code_skipped: 1
+      , (err, started) ->
+        return callback err if err
+        ctx.waitForConnection hive_metastore_host, hive_metastore_port, hive_metastore_timeout, (err) ->
+          callback err, started
   hive_metastore_stop: (ctx, callback) ->
     {hive_user, hive_pid_dir} = ctx.config.hdp
-    ctx.execute
-      # su -l hive -c "kill `cat /var/run/hive/metastore.pid`"
-      cmd: "su -l #{hive_user} -c \"kill `cat #{hive_pid_dir}/metastore.pid`\""
-      code_skipped: 1
-    , (err, started) ->
-      callback err, started
+    lifecyle.hive_metastore_status ctx, (err, running) ->
+      return callback err, false if err or not running
+      ctx.execute
+        # su -l hive -c "kill `cat /var/run/hive/metastore.pid`"
+        cmd: "su -l #{hive_user} -c \"kill `cat #{hive_pid_dir}/metastore.pid`\""
+        code_skipped: 1
+      , (err, started) ->
+        callback err, started
   hive_metastore_restart: (ctx, callback) ->
     lifecyle.hive_metastore_stop ctx, (err) ->
       return callback err if err
       lifecyle.hive_metastore_start ctx, callback
+  hive_server2_status: (ctx, callback) ->
+    # {hive_server2_pid} = ctx.config.hdp
+    lifecyle.is_pidfile_running ctx, "/var/run/hive/server2.pid", (err, running) ->
+      callback err, running
   hive_server2_start: (ctx, callback) ->
-    {hive_user, hive_log_dir, hive_pid_dir} = ctx.config.hdp
-    ctx.execute
-      # su -l hive -c 'nohup /usr/lib/hive/bin/hiveserver2 >/var/log/hive/hiveserver2.out 2>/var/log/hive/hiveserver2.log & echo $! >/var/run/hive/server2.pid'
-      # su -l hive -c '/usr/lib/hive/bin/hiveserver2'
-      cmd: "su -l #{hive_user} -c 'nohup /usr/lib/hive/bin/hiveserver2 >#{hive_log_dir}/hiveserver2.out 2>#{hive_log_dir}/hiveserver2.log & echo $! > /var/run/hive/server2.pid'"
-      code_skipped: 1
-    , (err, started) ->
-      callback err, started
+    {hive_user, hive_log_dir, hive_pid_dir, hive_server2_host, hive_server2_port, hive_server2_timeout} = ctx.config.hdp
+    lifecyle.hive_server2_status ctx, (err, running) ->
+      return callback err, false if err or running
+      ctx.execute
+        # su -l hive -c 'nohup /usr/lib/hive/bin/hiveserver2 >/var/log/hive/hiveserver2.out 2>/var/log/hive/hiveserver2.log & echo $! >/var/run/hive/server2.pid'
+        # su -l hive -c '/usr/lib/hive/bin/hiveserver2'
+        cmd: "su -l #{hive_user} -c 'nohup /usr/lib/hive/bin/hiveserver2 >#{hive_log_dir}/hiveserver2.out 2>#{hive_log_dir}/hiveserver2.log & echo $! > /var/run/hive/server2.pid'"
+        code_skipped: 1
+      , (err, started) ->
+        return callback err if err
+        ctx.waitForConnection hive_server2_host, hive_server2_port, hive_server2_timeout, (err) ->
+          callback err, started
   hive_server2_stop: (ctx, callback) ->
     {hive_user, hive_pid_dir} = ctx.config.hdp
-    ctx.execute
-      # su -l hive -c "kill `cat /var/run/hive/server2.pid"
-      cmd: "su -l #{hive_user} -c \"kill `cat #{hive_pid_dir}/server2.pid`\""
-      code_skipped: 1
-    , (err, started) ->
-      callback err, started
+    lifecyle.hive_server2_status ctx, (err, running) ->
+      return callback err, false if err or not running
+      ctx.execute
+        # su -l hive -c "kill `cat /var/run/hive/server2.pid"
+        cmd: "su -l #{hive_user} -c \"kill `cat #{hive_pid_dir}/server2.pid`\""
+        code_skipped: 1
+      , (err, started) ->
+        callback err, started
   hive_server2_restart: (ctx, callback) ->
     lifecyle.hive_server2_stop ctx, (err) ->
       return callback err if err
@@ -156,16 +176,6 @@ lifecyle = module.exports =
     {oozie_pid_dir} = ctx.config.hdp
     lifecyle.is_pidfile_running ctx, "#{oozie_pid_dir}/oozie.pid", (err, running) ->
       callback err, running
-    # ctx.execute
-    #   cmd: """
-    #   if pid=`cat #{oozie_pid_dir}/oozie.pid`; then
-    #     if ps cax | grep -v grep | grep $pid; then exit 0; else
-    #       rm -f #{oozie_pid_dir}/oozie.pid
-    #   fi; fi; exit 1
-    #   """
-    #   code_skipped: 1
-    # , (err, started) ->
-    #   callback err, started
   oozie_start: (ctx, callback) ->
     {oozie_user} = ctx.config.hdp
     lifecyle.oozie_status ctx, (err, running) ->
